Fxaa v1 is indeed done, and we run on msvc

+++ common/scene/ViewContext.hpp
@@ -63,9 +63,9 @@ private:
   const SceneManager& sceneMgr;
 
 private:
-  uint32_t indirectDrawBufByteSize() const { return sceneMgr.getIndirectCommands().size_bytes(); }
+  uint32_t indirectDrawBufByteSize() const { return uint32_t(sceneMgr.getIndirectCommands().size_bytes()); }
   uint32_t markedInstBufSizeBytes() const
   {
-    return sceneMgr.getInstances().size() * sizeof(DrawableInstance);
+    return uint32_t(sceneMgr.getInstances().size() * sizeof(DrawableInstance));
   }
 };
